Min fina knapp för att komma till inmatning (#61)
* Fixed input button, moved icons

* Works now when sidebar is collapsed, unsure if this is what we want tho

* Fixed input button, moved icons

* Works now when sidebar is collapsed, unsure if this is what we want tho

* Style check fix BROR

* Fixed so ul only has one selector
