--- frontend-friends/src/components/sidebar/icons/PlaceHolderIcon.js
@@ -1,15 +1,0 @@
-export default function StatisticsIcon () {
-  return (
-    <>
-      <svg width='24' height='24' viewBox='0 0 24 24' fill='none' xmlns='http://www.w3.org/2000/svg'>
-        <rect x='5' y='6' width='2' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='5' y='11' width='2' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='5' y='16' width='2' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='9' y='6' width='10' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='9' y='11' width='10' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='9' y='16' width='10' height='2' rx='0.5' fill='#579CFB' />
-        <rect x='1' y='1' width='22' height='22' rx='5' stroke='#579CFB' strokeWidth='2' />
-      </svg>
-    </>
-  )
-}
