+++ frontend-friends/src/components/sidebar/sidebar.scss
@@ -3,9 +3,8 @@
 .collapsed-sidebar .sidebar-container{
     width: 80px;
     .nav-text{
-        width: 0;
+        display: none;
         overflow: hidden;
-        height: 0;
     }
 
     .navigation-link{
@@ -13,7 +12,6 @@
         justify-content: center;
         padding: 0.5em 0.5em;
         gap: 0;
-        border-radius: 30px;
     }
 
     .default-logo{
@@ -25,7 +23,21 @@
     }
 
     button{
-        width: 0;
+        display: none;
+    }
+
+    .input-link{
+        .navigation-link{
+            width: 50px;
+
+            .nav-text{
+                display: none;
+            }
+
+            svg{
+                display: inline-block;
+            }
+        }
     }
 }
 
@@ -36,6 +48,7 @@
     background-color: #fff;
     width: 260px;
     height: 100%;
+    min-height: 370px;
     padding: 1em;
     position: fixed;
     z-index: 100;
@@ -64,7 +77,7 @@
         margin: 20px 0;
         padding: 0;
         display: flex;
-        align-items: center;
+        align-items: stretch;
         flex-direction: column;
 
         li{
@@ -91,7 +104,7 @@
         align-items: center;
         justify-content: flex-start;
         padding: 0.5em 1em;
-        border-radius: 1em;
+        border-radius: 6px;
         font-size: 20px;
         height: 28px;
         margin: 0.2em auto;
@@ -107,9 +120,40 @@
     }
 
     .input-link{
+        display: flex;
+        flex-direction: column;
+        margin: calc(1em - 7px) 0;
+
         .navigation-link{
             background-color: colors.$primary-color;
             color: white;
+            transition: background-color 0.4s;
+            height: 50px;
+            width: 220px;
+            padding: 0;
+            margin: 0;
+            align-self: center;
+
+            .nav-text{
+                display: flex;
+                align-items: center;
+                width: 100%;
+                justify-content: center;
+            }
+
+            svg{
+                display:none;
+            }
+
+            span{
+                svg{
+                    display: inline-block;
+
+                    path{
+                        fill: white;
+                    }
+                }
+            }
         }
 
         .navigation-link:hover{
@@ -121,11 +165,6 @@
     .active{
         color: colors.$primary-color;
     }
-
-    ul{
-        align-items: stretch;
-    }
-
 
     .nav-text{
         height: auto;
@@ -142,19 +181,19 @@
     }
     
     button {
-        all: unset;
         position:absolute;
         background-color: colors.$primary-color;
         font-weight: 600;
         color: white;
         text-align: center;
-        width: 70%;
+        width: 150px;
         height: 50px;
         border-radius: 6px;
         cursor: pointer;
         transition: 0.4s;
         bottom: 2%;
-        margin-left: 9%;
+        display:block;
+        align-self: center;
         
         &:hover {
             background-color: #4984D3;
@@ -168,9 +207,8 @@
     .sidebar-container{
         width: 80px;
         .nav-text{
-            width: 0;
+            display: none;
             overflow: hidden;
-            height: 0;
         }
     
         .navigation-link{
@@ -189,6 +227,20 @@
         }
         button{
             display: none;
+        }
+
+        .input-link{
+            .navigation-link{
+                width: 50px;
+
+                .nav-text{
+                    display: none;
+                }
+    
+                svg{
+                    display: inline-block;
+                }
+            }
         }
     }
