+++ frontend-friends/src/components/submitButton/submitButton.scss
@@ -11,6 +11,7 @@
     border-radius: 6px;
     cursor: pointer;
     transition: 0.4s;
+    text-transform: uppercase;
     
     &:hover {
         background-color: #4984D3;
