+++ frontend-friends/src/style/colors.scss
@@ -1,6 +1,6 @@
 $primary-color: #579CFB;
 $primary-dark: #4e8ce1;
-$light-grey: #F0F1F2;
+$light-grey: #EBEFF3;
 $darkerlight-grey: #e9ecf1;
 $grey-text: #A6A6A6;
 $dark-blue: #283055;
@@ -14,7 +14,7 @@ $orange: #ff922a;
 $yellow: #ffd43b;
 /* Category colors from Figma */
 $category-blue: $primary-color;
-$category-blue-dark: #3B6AAB;
+$category-blue-dark: $primary-dark;
 $category-cyan: #20E2BA;
 $category-cyan-dark: #169A7E;
 $category-red: $primary-red;
