+++ frontend-friends/src/utils/icons/ArrowForward.js
@@ -1,0 +1,7 @@
+export default function ArrowForward () {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' height='48' viewBox='0 96 960 960' width='48'>
+      <path d='m561 814-43-42 168-168H160v-60h526L517 375l43-42 241 241-240 240Z' />
+    </svg>
+  )
+}
