+++ frontend-friends/src/utils/icons/CheckIcon.js
