+++ frontend-friends/src/utils/icons/InputIcon.js
