+++ frontend-friends/src/utils/icons/MailIcon.js
