+++ frontend-friends/src/utils/icons/NotesIcon.js
