+++ frontend-friends/src/utils/icons/StatisticsIcon.js
